Fixed Routing and displayed Top Manga
Completed:
-Fixed the Top manga issue. It had to do with "/:anime" being above the
get request of "/getTopManga" so it got the "getTopManga" portion and
put it into the "/:anime" get request instead of the "/getTopManga".
To combat this, I added an extra part indicating what type of series it
is, ie: "/manga/getTopManga"
-Upon doing the top manga issue, I also changed the other routes so that
it would make more sense. There is still more room to change the routes.

In progress:
-Still need to finish the about page and displaying what tools are being
displayed.